refactor: change Integer to Long

+++ src/main/java/com/devcard/devcard/chat/repository/ChatRoomRepository.java
@@ -6,5 +6,5 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ChatRoomRepository extends JpaRepository<ChatRoom, Long> {
-    List<ChatUser> findByIdIn(List<Integer> participantsId);
+    List<ChatUser> findByIdIn(List<Long> participantsId);
 }
